requirements.txt
python-dotenv==1.0.1
rich==13.7.1
pandas==2.2.2
